segregate views_profile from views_auth
