+++ oga/backend/users/views/__init__.py
@@ -5,3 +5,4 @@ from .views_notification import *
 from .views_location import *
 from .views_answer import *
 from .views_rate import *
+from .views_profile import *
