+++ oga/backend/users/views/views_auth.py
@@ -68,33 +68,3 @@ def logged_out(request):
     """
     logout(request)
     return JsonResponse({}, status=204)
-
-
-@check_login_required
-@ensure_csrf_cookie
-@require_http_methods(["GET"])
-def get_profile(request, username=''):
-    """
-    get profile details of a user based on username
-    """
-    if username != "":
-        # get profile details of currently logged in user
-        user = User.objects.get(username=username)
-    else:
-        user = get_user(request)
-    profile = Profile.objects.get(user=user)
-    location_name = "Location unknown"
-    coordinates = ""
-
-    if profile.location_id:
-        location_name = profile.location_id.name
-        coordinates = "(" + str(round(profile.location_id.latitude, 2)) + \
-            ", " + str(round(profile.location_id.longitude, 2)) + ")"
-
-    response_dict = {
-        'id': profile.id,
-        'username': profile.user.username,
-        'location': location_name,
-        'coordinates': coordinates
-    }
-    return JsonResponse(response_dict, status=200)
